+++ src/Microservices/AccountsMS/AccountsMS.Business/Services/Implementation/OrganizationService.cs
@@ -1,5 +1,6 @@
 ﻿using AccountsMS.Business.DTOs.Organization;
-using AccountsMS.Business.Exceptions;
+using AccountsMS.Business.Responce.Enums;
+using AccountsMS.Business.Responce.NonGeneric;
 using AccountsMS.Business.Services.Contracts;
 using AccountsMS.Data.Models.Organization;
 using AccountsMS.Data.Repositories.Contracts;
@@ -18,40 +19,48 @@ namespace AccountsMS.Business.Services.Implementation
             _mapper = mapper;
         }
 
-        public async Task<IEnumerable<OrganizationViewDto>> GetAllOrganizationsAsync()
+        public async Task<Result> GetAllOrganizationsAsync()
         {
             var organizationsModels = await _organizationRepository.GetAllOrganizationsAsync();
 
             var organizationsViewDtos = _mapper.Map<IEnumerable<OrganizationViewDto>>(organizationsModels);
 
-            return organizationsViewDtos;
+            return Result<IEnumerable<OrganizationViewDto>>.Successed(organizationsViewDtos);
         }
 
-        public async Task<OrganizationViewDto> GetOrganizationByIdAsync(int id)
+        public async Task<Result> GetOrganizationByIdAsync(int id)
         {
-            var organizationModel = await _organizationRepository.GetOrganizationByIdAsync(id)
-                ?? throw new NotFoundException("Organization was not found.");
+            var organizationModel = await _organizationRepository.GetOrganizationByIdAsync(id);
+
+            if (organizationModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Organization not found.");
+            }
 
             var organizationViewDto = _mapper.Map<OrganizationViewDto>(organizationModel);
 
-            return organizationViewDto;
+            return Result<OrganizationViewDto>.Successed(organizationViewDto);
         }
 
-        public async Task<OrganizationViewDto> GetOrganizationByEmailAsync(string email)
+        public async Task<Result> GetOrganizationByEmailAsync(string email)
         {
-            var organizationModel = await _organizationRepository.GetOrganizationByEmailAsync(email)
-                ?? throw new NotFoundException("Organization was not found.");
+            var organizationModel = await _organizationRepository.GetOrganizationByEmailAsync(email);
+
+            if (organizationModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Organization not found.");
+            }
 
             var organizationViewDto = _mapper.Map<OrganizationViewDto>(organizationModel);
 
-            return organizationViewDto;
+            return Result<OrganizationViewDto>.Successed(organizationViewDto);
         }
 
-        public async Task<OrganizationViewDto> CreateOrganizationAsync(OrganizationCreateDto organizationCreateDto)
+        public async Task<Result> CreateOrganizationAsync(OrganizationCreateDto organizationCreateDto)
         {
             if (organizationCreateDto == null)
             {
-                throw new ArgumentNullException(nameof(organizationCreateDto));
+                return Result.Failed(AccountsMSErrorCodes.NullArgument, nameof(organizationCreateDto));
             }
 
             var organizationCreateModel = _mapper.Map<OrganizationCreateModel>(organizationCreateDto);
@@ -60,30 +69,42 @@ namespace AccountsMS.Business.Services.Implementation
 
             var organizationViewDto = _mapper.Map<OrganizationViewDto>(organizationModel);
 
-            return organizationViewDto;
+            return Result<OrganizationViewDto>.Successed(organizationViewDto);
         }
 
-        public async Task UpdateOrganizationAsync(OrganizationUpdateDto organizationUpdateDto)
+        public async Task<Result> UpdateOrganizationAsync(OrganizationUpdateDto organizationUpdateDto)
         {
             if (organizationUpdateDto == null)
             {
-                throw new ArgumentNullException(nameof(organizationUpdateDto));
+                return Result.Failed(AccountsMSErrorCodes.NullArgument, nameof(organizationUpdateDto));
             }
 
-            var existingOrganizationModel = await _organizationRepository.GetOrganizationByIdAsync(organizationUpdateDto.Id)
-                ?? throw new NotFoundException("Organization was not found.");
+            var existingOrganizationModel = await _organizationRepository.GetOrganizationByIdAsync(organizationUpdateDto.Id);
+
+            if (existingOrganizationModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Organization not found.");
+            }
 
             var organizationUpdateModel = _mapper.Map<OrganizationUpdateModel>(organizationUpdateDto);
 
             await _organizationRepository.UpdateOrganizationAsync(organizationUpdateModel);
+
+            return Result.Successed();
         }
 
-        public async Task DeleteOrganizationAsync(int id)
+        public async Task<Result> DeleteOrganizationAsync(int id)
         {
-            var existingOrganizationModel = await _organizationRepository.GetOrganizationByIdAsync(id)
-                ?? throw new NotFoundException("Organization was not found.");
+            var existingOrganizationModel = await _organizationRepository.GetOrganizationByIdAsync(id);
+
+            if (existingOrganizationModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Organization not found.");
+            }
 
             await _organizationRepository.DeleteOrganizationAsync(id);
+
+            return Result.Successed();
         }
     }
 }
