+++ src/Microservices/AccountsMS/AccountsMS.Business/Services/Implementation/SkillService.cs
@@ -1,5 +1,7 @@
 ﻿using AccountsMS.Business.DTOs.Skill;
 using AccountsMS.Business.Exceptions;
+using AccountsMS.Business.Responce.Enums;
+using AccountsMS.Business.Responce.NonGeneric;
 using AccountsMS.Business.Services.Contracts;
 using AccountsMS.Data.Models.Skill;
 using AccountsMS.Data.Repositories.Contracts;
@@ -18,30 +20,34 @@ namespace AccountsMS.Business.Services.Implementation
             _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
-        public async Task<IEnumerable<SkillViewDto>> GetAllSkillsAsync()
+        public async Task<Result> GetAllSkillsAsync()
         {
             var skillsModels = await _skillRepository.GetAllSkillsAsync();
 
             var skillsViewDtos = _mapper.Map<IEnumerable<SkillViewDto>>(skillsModels);
 
-            return skillsViewDtos;
+            return Result<IEnumerable<SkillViewDto>>.Successed(skillsViewDtos);
         }
 
-        public async Task<SkillViewDto> GetSkillByIdAsync(int id)
+        public async Task<Result> GetSkillByIdAsync(int id)
         {
-            var skillModel = await _skillRepository.GetSkillByIdAsync(id)
-                ?? throw new NotFoundException("Skill was not found.");
+            var skillModel = await _skillRepository.GetSkillByIdAsync(id);
+
+            if(skillModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Skill not found.");
+            }
 
             var skillViewDto = _mapper.Map<SkillViewDto>(skillModel);
 
-            return skillViewDto;
+            return Result<SkillViewDto>.Successed(skillViewDto);
         }
 
-        public async Task<SkillViewDto> CreateSkillAsync(SkillCreateDto skillCreateDto)
+        public async Task<Result> CreateSkillAsync(SkillCreateDto skillCreateDto)
         {
             if (skillCreateDto == null)
             {
-                throw new ArgumentNullException(nameof(skillCreateDto));
+                return Result.Failed(AccountsMSErrorCodes.NullArgument, nameof(skillCreateDto));
             }
 
             var skillCreateModel = _mapper.Map<SkillCreateModel>(skillCreateDto);
@@ -50,30 +56,42 @@ namespace AccountsMS.Business.Services.Implementation
 
             var createdSkillViewDto = _mapper.Map<SkillViewDto>(createdSkillModel);
 
-            return createdSkillViewDto;
+            return Result<SkillViewDto>.Successed(createdSkillViewDto);
         }
 
-        public async Task UpdateSkillAsync(SkillUpdateDto skillUpdateDto)
+        public async Task<Result> UpdateSkillAsync(SkillUpdateDto skillUpdateDto)
         {
             if (skillUpdateDto == null)
             {
-                throw new ArgumentNullException(nameof(skillUpdateDto));
+                return Result.Failed(AccountsMSErrorCodes.NullArgument, nameof(skillUpdateDto));
             }
 
-            var existingSkillModel = await _skillRepository.GetSkillByIdAsync(skillUpdateDto.Id)
-                ?? throw new NotFoundException("Skill was not found.");
+            var existingSkillModel = await _skillRepository.GetSkillByIdAsync(skillUpdateDto.Id);
+
+            if (existingSkillModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Skill not found.");
+            }
 
             var skillUpdateModel = _mapper.Map<SkillUpdateModel>(skillUpdateDto);
 
             await _skillRepository.UpdateSkillAsync(skillUpdateModel);
+
+            return Result.Successed();
         }
 
-        public async Task DeleteSkillAsync(int id)
+        public async Task<Result> DeleteSkillAsync(int id)
         {
-            var existingSkillModel = await _skillRepository.GetSkillByIdAsync(id)
-                ?? throw new NotFoundException("Skill was not found.");
+            var existingSkillModel = await _skillRepository.GetSkillByIdAsync(id);
+
+            if (existingSkillModel == null)
+            {
+                return Result.Failed(AccountsMSErrorCodes.EntityNotFound, "Skill not found.");
+            }
 
             await _skillRepository.DeleteSkillAsync(id);
+
+            return Result.Successed();
         }
     }
 }
